organizing project for publication
